services/orchestrator/requirements.txt
google-adk>=0.5.0
google-cloud-aiplatform>=1.50.0
vertexai>=1.50.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0
structlog>=24.1.0
